pyproject.toml
[project]
name = "polyclaw"
version = "0.1.0"
description = "Trading-enabled Polymarket skill for OpenClaw"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "web3>=7.0.0",
    "httpx[socks]>=0.28.0",
    "py-clob-client>=0.34.0",
    "eth-account>=0.13.0",
    "python-dotenv>=1.0.0",
    "python-telegram-bot>=21.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["lib", "scripts"]
